Converting temperature model to temperature entity model

+++ src/EcleticaBeerControl.Worker/Payloads/Temperature.cs
@@ -4,7 +4,7 @@ namespace EcleticaBeerControl.Worker.Payloads
 {
     public sealed record Temperature
     {
-        public double Degrees { get; set; }
+        public long Degrees { get; set; }
         public DegreeFormat DegreeFormat { get; set; }
     }
 }
